Convert tstring to submodule
