+++ tsdat/tstring/get_regex_for_var.py
@@ -1,0 +1,21 @@
+def _get_regex_for_var(var_name: str) -> str:
+    return dict(
+        year=r"(?P<year>[0-9]{4})",
+        month=r"(?P<month>[0-9]{2})",
+        day=r"(?P<day>[0-9]{2})",
+        hour=r"(?P<hour>[0-9]{2})",
+        minute=r"(?P<minute>[0-9]{2})",
+        second=r"(?P<second>[0-9]{2})",
+        yyyy=r"(?P<yyyy>[0-9]{4})",
+        mm=r"(?P<mm>[0-9]{2})",
+        dd=r"(?P<dd>[0-9]{2})",
+        HH=r"(?P<HH>[0-9]{2})",
+        MM=r"(?P<MM>[0-9]{2})",
+        SS=r"(?P<SS>[0-9]{2})",
+        date_time=r"(?P<date_time>[0-9]{8}\.[0-9]{6})",
+        date=r"(?P<date>[0-9]{8})",
+        time=r"(?P<time>[0-9]{6})",
+        start_date=r"(?P<start_date>[0-9]{8})",
+        start_time=r"(?P<start_time>[0-9]{6})",
+        datastream=r"(?P<datastream>[a-zA-Z0-9.-_]+)",
+    ).get(var_name, f"(?P<{var_name}>[a-zA-Z0-9]+)")
