+++ tsdat/tstring/template.py
@@ -4,128 +4,8 @@ import re
 from pathlib import Path
 from typing import Callable, Mapping
 
-__all__ = ("Template",)
-
-
-def _is_balanced(template: str) -> bool:
-    stack: list[str] = []
-    for char in template:
-        if char in "{[":
-            stack.append("}" if char == "{" else "]")
-        elif char in "}]":
-            if not stack or char != stack.pop():
-                return False
-    return len(stack) == 0
-
-
-def _get_regex_for_var(var_name: str) -> str:
-    return dict(
-        year=r"(?P<year>[0-9]{4})",
-        month=r"(?P<month>[0-9]{2})",
-        day=r"(?P<day>[0-9]{2})",
-        hour=r"(?P<hour>[0-9]{2})",
-        minute=r"(?P<minute>[0-9]{2})",
-        second=r"(?P<second>[0-9]{2})",
-        yyyy=r"(?P<yyyy>[0-9]{4})",
-        mm=r"(?P<mm>[0-9]{2})",
-        dd=r"(?P<dd>[0-9]{2})",
-        HH=r"(?P<HH>[0-9]{2})",
-        MM=r"(?P<MM>[0-9]{2})",
-        SS=r"(?P<SS>[0-9]{2})",
-        date_time=r"(?P<date_time>[0-9]{8}\.[0-9]{6})",
-        date=r"(?P<date>[0-9]{8})",
-        time=r"(?P<time>[0-9]{6})",
-        start_date=r"(?P<start_date>[0-9]{8})",
-        start_time=r"(?P<start_time>[0-9]{6})",
-        datastream=r"(?P<datastream>[a-zA-Z0-9.-_]+)",
-    ).get(var_name, f"(?P<{var_name}>[a-zA-Z0-9]+)")
-
-
-class TemplateChunk:
-    """Class to hold a chunk of a Template.
-
-    A chunk is a smaller part of a template. E.g., the template "{a}{b}{c}d" has 4
-    chunks: "{a}", "{b}", "{c}", and "d"."""
-
-    def __init__(self, chunk: str) -> None:
-        if not _is_balanced(chunk):
-            raise ValueError(f"Unbalanced brackets in chunk: '{chunk}'")
-        self.str = chunk
-        self.var_name = self._get_variable_name(chunk)
-        self.parts = self._get_parts(chunk)
-        self.is_required = self._is_required(chunk)
-        self.regex = self._generate_regex(chunk)
-
-    def __repr__(self) -> str:
-        return f"TemplateChunk({self.str})"
-
-    @staticmethod
-    def _get_variable_name(chunk: str) -> str | None:
-        if "{" in chunk:
-            start, stop = chunk.index("{"), chunk.index("}")
-            return chunk[start + 1 : stop]
-        return None
-
-    @staticmethod
-    def _get_parts(chunk: str) -> tuple[str, ...]:
-        return tuple(re.split(r"[{}\[\]]", chunk))
-
-    @staticmethod
-    def _is_required(chunk: str) -> bool | None:
-        if "{" in chunk:
-            return "[" not in chunk
-        return None
-
-    @staticmethod
-    def _generate_regex(chunk: str) -> str:
-        regex_pattern = ""
-        i = 0
-        while i < len(chunk):
-            char = chunk[i]
-            if char == "{":
-                var_start = i + 1
-                var_end = chunk.index("}", var_start)
-                var_name = chunk[var_start:var_end]
-                regex_pattern += _get_regex_for_var(var_name)
-                i = var_end + 1
-            elif char == "[":
-                regex_pattern += "(?:"
-                i += 1
-            elif char == "]":
-                regex_pattern += ")?"
-                i += 1
-            else:
-                regex_pattern += re.escape(char)
-                i += 1
-        return regex_pattern
-
-    def sub(
-        self,
-        value: str | (Callable[[], str]) | None,
-        allow_missing: bool = False,
-        fill: str | None = None,
-    ) -> str:
-        def remove_square_brackets(s: str) -> str:
-            return s if self.is_required else s[1:-1]
-
-        if callable(value):
-            value = value()
-
-        result = ""
-        if self.var_name is None:
-            result = self.str
-        elif value is not None:
-            result = self.str.replace(f"{{{self.var_name}}}", value)
-            result = remove_square_brackets(result)
-        elif allow_missing and fill:
-            result = fill
-        elif allow_missing:
-            result = self.str
-        elif self.is_required:
-            raise ValueError(
-                f"Could not make substitution for {self.var_name} in {self}"
-            )
-        return result
+from .is_balanced import _is_balanced
+from .template_chunk import TemplateChunk
 
 
 class Template:
